requirements.txt
# Web framework
fastapi>=0.104.0
httpx>=0.27.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# PDF parsing
pdfplumber>=0.10.0

# OCR (тянет PyTorch ~2 ГБ при первой установке)
easyocr>=1.7.0

# Image processing
Pillow>=10.0.0

# Session storage (опционально — fallback на in-memory при отсутствии Redis)
redis>=5.0.0

# Data validation
pydantic>=2.0.0
